pyproject.toml
[tool.poetry]
name = "github-practice-evaluation"
version = "0.1.0"
description = "Evaluate GitHub Practices"
authors = ["Hansae Ju <sparky@hanyang.ac.kr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
icecream = "^2.1.3"
rich = "^13.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
